+++ tests/cli_pl.rs
@@ -1,0 +1,64 @@
+use assert_cmd::prelude::*; // Add methods on commands
+use predicates::prelude::*; // Used for writing assertions
+use std::process::Command; // Run programs
+
+#[test]
+fn command_overlap() -> anyhow::Result<()> {
+    let mut bin = String::new();
+    for e in &["LAshow"] {
+        if let Ok(pth) = which::which(e) {
+            bin = pth.to_string_lossy().to_string();
+            break;
+        }
+    }
+    if bin.is_empty() {
+        return Ok(());
+    } else {
+        eprintln!("bin = {:#?}", bin);
+    }
+
+    let mut cmd = Command::cargo_bin("anchr")?;
+    let output = cmd
+        .arg("overlap")
+        .arg("tests/ovlpr/1_4.pac.fasta")
+        .output()
+        .unwrap();
+    let stdout = String::from_utf8(output.stdout).unwrap();
+
+    assert_eq!(stdout.lines().count(), 18);
+    assert!(stdout.contains("pac4745_7148"), "original names");
+    assert!(stdout.contains("contained"), "relations");
+
+    let mut cmd = Command::cargo_bin("anchr")?;
+    let output = cmd
+        .arg("overlap")
+        .arg("--idt")
+        .arg("0.8")
+        .arg("--len")
+        .arg("2500")
+        .arg("--serial")
+        .arg("tests/ovlpr/1_4.pac.fasta")
+        .output()
+        .unwrap();
+    let stdout = String::from_utf8(output.stdout).unwrap();
+
+    assert_eq!(stdout.lines().count(), 4);
+    assert!(!stdout.contains("pac4745_7148"), "serials");
+
+    let mut cmd = Command::cargo_bin("anchr")?;
+    let output = cmd
+        .arg("overlap")
+        .arg("--idt")
+        .arg("0.8")
+        .arg("--len")
+        .arg("2500")
+        .arg("--all")
+        .arg("tests/ovlpr/1_4.pac.fasta")
+        .output()
+        .unwrap();
+    let stdout = String::from_utf8(output.stdout).unwrap();
+
+    assert_eq!(stdout.lines().count(), 42);
+
+    Ok(())
+}